refactor(download): Introduce Downloadable interface and use with buckets
This is the first of a series of PRs with the goal of making the download code - and later also the deploy and delete code - independent of the concrete feature/API.
We introduce a `Downloadable` interface and use for already for downloading buckets. The goal is to use it for all the other clients as well. As soon as all clients adhere to `Downloadable`, we can refactor/simplify `download_configs.go`