refactor: move deployment limiter into context (#1841)
#### What this PR does / Why we need it:
Removes the global variable and puts deployment options into the
context, to prevent prop-drilling

#### Special notes for your reviewer:

Coverage check may fail because of these two lines inside the `if`
```go
if options.ConcurrentDeploymentsLimiter != nil {
	options.ConcurrentDeploymentsLimiter.Acquire()
	defer options.ConcurrentDeploymentsLimiter.Release()
}
```
The question is: do we want to test that?
Add a test that just checks if acquire/release was called? how: replace
the limiter struct inside the options with an interface and put a stub
into the context, so that the call can be checked.
Should we write an integration test for it? how: stubbed clients that
don't respond and we check if upserts were called
{limitedConcurrentCount} times

#### Does this PR introduce a user-facing change?
No