refactor(download): Refactor automation download to use Downloadable interface
This is one PR in a series of PRs with the goal of making the download code - and later also the deploy and delete code - independent of the concrete resource api.
The `Downloadable` interface, is used for downloading automation resources. As soon as all clients adhere to `Downloadable`, we can refactor/simplify `download_configs.go`