refactor(settings): settings client uses permission handling from core lib
This commit removes all the types and most of the logic concerning interactions with the settings object permission API from settings_client.go. This code is moved to the core library, such that it can be used also in the Terraform provider.

Consequently, all references to the now-removed dtclient.PermissionObject and nested types are updated in this commit.

It was decided that the Upsert is not moved. Instead, it uses the newly written Create and Update functions from the core library.

When creating a new settings client, the nested rest client is also used to create the core lib permission client.