ci: remove additional reporting and pipeline fails for linting
Because Sonar does its own reporting, we actually don't need our lint action to report (and fail) an additional time. Therefore, it's enough if the lint report is generated, which is everything Sonar needs. This also adds the ability to ignore specific errors that are reported not to occur again, once ignored