refactor: use different endpoint for classic connection check (#1862)
#### What this PR does / Why we need it:
For validating a classic client, the cluster version was retrieved in
order to check if the connection is actually working.
The downside of this is, that an additional scope (`DataExport`) is
needed for this, which is not optimal.
The connection check is now changed to use the token lookup endpoint.
This one doesn't require any specific scopes.
`DataExport` is not required anymore but still optional, as this is
still needed for checking deprecated cluster versions that are not
compatible with settings.

#### Does this PR introduce a user-facing change?
The `DataExport` scope is not required anymore

---------

Co-authored-by: Arthur Pitman <arthur.pitman@dynatrace.com>