refactor(download): downloadFn is removed from download_configs.go
A `prepareDownloadables` function is introduced, which constructs all the resource APIs, based on the `ClientSet` and the `downloadConfigsOptions`. In `downloadConfigs`, this factory function is called, we iterate over all `Downloadable` instances, call `Download`, and collect and return the resulting configs. This is in contrast to the previous implementation, where constructing the "client" and downloading the configs all happened in one go. The new implementation would allow for parallel downloading, if we'd wish to do that.

For `TestDownload_Options`, instead of checking, which download functions have been called on `downloadConfigs`, we check, which Downloadable instances have been created by `prepareDownloadables`