refactor(download): Unify `downloadConfigsOptions` creation
There are two functions in `download_configs.go` which create the `downloadConfigsOptions` struct: `DownloadConfigsBasedOnManifest` and `DownloadConfigs`. The code is nearly the same for both cases, so I extracted the function `toDownloadConfigsOptions` to have creation of the struct at one place. This function is called by the two formerly mentioned functions.