refactor(config): use retry logic of core (#1895)
#### **Why** this PR?
The config client had its own retry implementation, leading to
duplicated code and more maintenance if a change is needed on both ends.
Now it's using the retry logic of core and not using a custom one
anymore.

#### **What** has changed?
The config client now uses the retry mechanism of core instead of having
its own.

#### **How** does it do it?
It's now using the retry functionality of core.

#### How is it **tested**?
A new test was added that tests the retry behavior

#### How does it affect **users**?
It slightly changes the logs, but nothing else