feat(settings): Remove server version check
For the settings client, it was important to check whether the Dynatrace server version is at least 262, as servers with versions < 1.262 are not able to handle updates of existing settings 2.0 objects that are non-deletable.

Since 1.262 is relative old now and not supported any longer, we assume this check to be obsolete.