feat(clients): parses and uses additional HTTP headers
The additional http headers set via `MONACO_ADDITIONAL_HTTP_HEADERS` are parsed and used in all places where a client factory is used to create REST clients. Additionally, it is parsed once in the runner in order to log the keys of the additional HTTP headers right at the start of Monaco.