fix: adding -p 1 to integration test execution commands.
We are facing problems when executing end-to-end tests. In particular, "unexpected EOF" errors occur in almost every execution. This hints at a deeper problem in our handling of HTTP requests, which we did not find yet. The problem started to appear after the last end-to-end test refactoring where one large test package was split up into several smaller ones. Therefore, the opportunity to execute multiple test packages in parallel presented itself. This caused the EOF problem to surface. Adding the "-p 1" command line option to the test execution cause all test packages to be executing serially. This effectively hides the problem, but doesn't fix it. Therefore, this is a TEMPORARY crutch which ought to be reverted again, as soon as the root cause is found and fixed.