refactor(classic): classic delete follows interface
For this, I needed to split the `Deleter` interface. The `DeleteAll` comes now from another interface, named `Purger`. For the classic configs, there are two components: A `Deleter` component in `delete.go` and a `Purger` component in newly created `purge.go`. The reason is that the `DeleteAll` function had a different signature for classic configs. Instead of deleting ALL configs, you could pass a set of specific APIs which were to be purged. Configs of APIs not included there were not touched. It made no sense to change the signature of `DeleteAll` in the `Deleter` interface just for that.

The solution I chose for handling this is the introduction of a new `Purger` interface, providing `DeleteAll`. It would still not make any sense to include the `api.APIs` argument in the signature of the function. I chose to add it to the `classic.Purger` component instead.