fix(logging): removing warning logs for unknown env overrides
We need to stop printing a warning log for every config which has an environment override for an environment that we do not know. The reason is that in practice, you might have a manifest with multiple environments defined, and environment overrides for some of them, but you might use this manifest in a deploy-command and only specifying a subset of environments to deploy to, using the "-e <ENV>" option. If you then have a lot of configs with environment overrides for an environment != <ENV>, you will get lots of warning logs. Effectively, this is log spam, making log analysis much harder.