fix: `TeeHandler` should call `Handler` on  a dependent  handler only if it is enabled for the given record
On closer inspection, `Handler.Handle()` should only be called if the `Handler` is enabled for the given record. This fixes that for the `TeeHandler` checking this for each dependent handler.

Note, the implementation of `Enabled` is still valid, and prevents `Handle` from even being called in cases where no dependent handler is enabled for the given record.

This fix is important as without it, all logs end up in the error log file.