refactor(bucket): use core lib without upsert and active check (#1969)
#### **Why** this PR?
The logic for updating a bucket only when it can be is now handled by
Monaco and not the core lib. The core lib also removes upsert, which
Monaco depends on.

#### **What** has changed?
- The logic for updating a bucket only when it can be, is now handled by
Monaco and not the core lib
- The upsert was removed, so Monaco now handles the upsert
(create/update) logic.

#### **How** does it do it?
By splitting Upsert into create & update, and adding some checks before
it can be updated/deleted.

#### How is it **tested**?
New tests are added, existing ones are adapted, and some existing ones
are moved closer to the package that is being tested.

#### How does it affect **users**?
NONE

#### Core changes

https://github.com/Dynatrace/dynatrace-configuration-as-code-core/pull/235

### Notes
Only realized later that there were adjustments of other PRs pointing to
the same files. Anyway, a bit of refactoring doesn't hurt I guess 😉

**Issue:** CA-15100