fix: only parse account or environments envs depending on the usage (#1982)
#### **Why** this PR?
During `monaco account deploy`, the envs of the environments should not
be validated, and for `monaco deploy`, the envs of the accounts should
not be validated

#### **What** has changed?
- Env vars of environmentGroups are not validated when using account
commands
- Env vars of account are not validated when not using account commands

#### **How** does it do it?
By excluding the parsing of env vars during manifest load.

#### How is it **tested**?
Existing tests cover it and new ones were added that check if invalid
envs can and can't be set.

#### How does it affect **users**?
They can now use account commands without the need of providing the envs
for non-account related commands and vice versa.

**Issue:** CA-15587