feat(settings): support admin mode for permissions (#1992)
#### **Why** this PR?
The core lib added some functionality for setting the `adminAccess`
property for permissions. As this is breaking our current
implementation, we need to update the logic to handle the `adminAccess`
correctly.

#### **What** has changed?
- Updated the core lib
- Added a retry mechanism if `adminAccess` doesn't work

#### **How** does it do it?

#### How is it **tested**?
New tests added that check if the property is correctly set and the
retry correctly works.

#### How does it affect **users**?
No change for the users, as the settings client doesn't support the
`adminAccess` so far.

**Issue:** CA-15757

## Open for discussion
Should I move the `ExecuteWithAdminAccess` utility function to core?