feat(boundary): Introduces boundaries in bindings
When assigning policies to a group on account or environment level, boundaries can now be assigned in this binding as well, but only if the policy reference is nested within a "policy" element.

This structure is reflected in `resources.go` and `types.go`. Also, `types.go` features custom unmarshalling for this type, making also sure that the old way of defining policy bindings is still valid, as long as no boundaries are assigned.

Also featured is the writing of boundaries and policy bindings featuring boundaries.

As logic in deployer, downloader and validation depends on the structure which is now changed, minimal changes very made in `deployer.go`, `groups.go` and some test files. Examples have been provided also for `load_test.go`