fix: Only generate references to documents if ID is prefixed by a forward slash
With recent changes to the documents API, custom IDs may be used with little restriction. This effectively broke the generation of references to documents as these IDs often occurred in other situations. As a workaround, references are now only generated if the ID is prefixed by a forward slash, with the intent of only creating references in URLs pointing to documents. This is the default behavior; however this restriction can be removed by setting the feature flag `MONACO_RESTRICT_DOCUMENT_REFERENCE_CREATION` to `false`.