fix(support-archive): write command errors
The error returned from the command was not inside the support archive. That was because the errors returned inside the command run weren't logged inside, and the archive is written right after the run via the onFinalize of the command. Therefore, writing the support archive happened before the returned error was handled/logged by the parent that executed the command.