fix: potential panic while downloading users
If a user is downloaded, it may be the case that the user has new groups assigned, which are not in the group-configs fetched before. This resulted in adding `nil` instead of an `account.Ref` to an array, which panicked at some point when calling `ref.ID()`. To fix that, we only assign a reference if possible (aligned with policies and boundaries)