feat(credentials): Endpoint for CredentialVault config changes from Config API to Environment API
The credentials Config API has been deprecated for a long time now, and we want to switch to the Environment API for credentials. Since the calls and payloads for both the ConfigV1 and the EnvironmentV2 endpoint are identical, credentials created via the ConfigV1 endpoint also get returned via the EnvironmentV2 endpoint, and vice versa, we can just change the endpoint. The only difference is that the EnvironmentV2 GET endpoint also supports pagination, but we handle that already in the classic download client.

To test this, I, amongst other tests, used a monaco config to create credentials via the ConfigV1 endpoint, modified them via the EnvironmentV2 endpoint and deleted them again via the EnvironmentV2 endpoint. Then I did the same thing the other way around. Worked without issues.